can-visualization/requirements.txt
python-can>=4.0.0
cantools>=37.0.0
PyQt5>=5.15.0
pyqtgraph>=0.12.0